+++ plugin/src/main/kotlin/ru/cian/rustore/publish/RustorePublishCli.kt
@@ -12,6 +12,7 @@ internal data class RustorePublishCli(
     val releasePhasePercent: String? = null,
     val releaseNotes: String? = null,
     val apiStub: Boolean? = null,
-    val seoTags: List<SeoTag>? = null,
+    val seoTagIds: List<Int>? = null,
     val minAndroidVersion: String? = null,
+    val appType: AppTypes? = null,
 )