+++ plugin/src/main/kotlin/ru/cian/rustore/publish/RustorePublishTask.kt
@@ -135,12 +135,12 @@ abstract class RustorePublishTask : DefaultTask() {
     @SuppressWarnings("MaxLineLength")
     @get:Internal
     @set:Option(
-        option = "seoTags",
-        description = "List of release SEO tags from ru.cian.rustore.publish.SeoTag. " +
+        option = "seoTagIds",
+        description = "List of release SEO tag ids. " +
             "Number of tags should not be greater than 5. " +
             "For more details see documentation: https://www.rustore.ru/help/work-with-rustore-api/api-upload-publication-app/app-tag-list"
     )
-    var seoTags: String? = null
+    var seoTagIds: String? = null
 
     @SuppressWarnings("MaxLineLength")
     @get:Internal
@@ -152,6 +152,16 @@ abstract class RustorePublishTask : DefaultTask() {
             "Actual available values see on https://www.rustore.ru/help/work-with-rustore-api/api-upload-publication-app/create-draft-version"
     )
     var minAndroidVersion: String? = null
+
+    @SuppressWarnings("MaxLineLength")
+    @get:Internal
+    @set:Option(
+        option = "appType",
+        description = "Type of application version. Available values: [\"MAIN\", \"GAMES\"]. " +
+            "'MAIN' for non-game applications, 'GAMES' for games. " +
+            "See https://www.rustore.ru/help/work-with-rustore-api/api-upload-publication-app/create-draft-version"
+    )
+    var appType: AppTypes? = null
 
     @get:Internal
     @set:Option(option = "apiStub", description = "Use RestAPI stub instead of real RestAPI requests")
@@ -174,9 +184,10 @@ abstract class RustorePublishTask : DefaultTask() {
             buildFile = buildFile,
             releasePhasePercent = releasePhasePercent,
             releaseNotes = releaseNotes,
-            seoTags = seoTags?.split(",")?.map { SeoTag.valueOf(it.trim()) },
+            seoTagIds = seoTagIds?.split(",")?.map { it.trim().toInt() },
             apiStub = apiStub,
             minAndroidVersion = minAndroidVersion,
+            appType = appType,
         )
 
         rustoreLogger.i("extension=$extension")
@@ -234,9 +245,10 @@ abstract class RustorePublishTask : DefaultTask() {
             applicationId = config.applicationId,
             whatsNew = config.releaseNotes?.first()?.newFeatures ?: "",
             publishType = config.publishType.name,
-            seoTags = config.seoTags.take(5).map { it.id },
+            seoTagIds = config.seoTagIds.take(5),
             minAndroidVersion = config.minAndroidVersion,
             developerContacts = config.developerContacts,
+            appType = config.appType?.value,
         )
 
         rustoreLogger.v("5/6. Upload build file '${config.artifactFile}'")