+++ plugin/src/main/kotlin/ru/cian/rustore/publish/models/request/AppDraftRequest.kt
@@ -8,9 +8,11 @@ internal data class AppDraftRequest(
     @SerializedName("publishType")
     val publishType: String,
     @SerializedName("seoTagIds")
-    val seoTags: List<Int>,
+    val seoTagIds: List<Int>,
     @SerializedName("minAndroidVersion")
     val minAndroidVersion: String,
     @SerializedName("developerContacts")
     val developerContacts: DeveloperContactsRequestModel,
+    @SerializedName("appType")
+    val appType: String? = null,
 )