+++ plugin/src/main/kotlin/ru/cian/rustore/publish/service/RustoreService.kt
@@ -21,9 +21,10 @@ internal interface RustoreService {
         applicationId: String,
         whatsNew: String,
         publishType: String,
-        seoTags: List<Int>,
+        seoTagIds: List<Int>,
         minAndroidVersion: String,
         developerContacts: DeveloperContactsConfig,
+        appType: String?,
     ): Int
 
     @Suppress("LongParameterList")