+++ plugin/src/main/kotlin/ru/cian/rustore/publish/utils/ConfigProvider.kt
@@ -8,7 +8,6 @@ import ru.cian.rustore.publish.PluginConfig
 import ru.cian.rustore.publish.ReleaseNotesConfig
 import ru.cian.rustore.publish.ReleasePhaseConfig
 import ru.cian.rustore.publish.RustorePublishExtensionConfig
-import ru.cian.rustore.publish.SeoTag
 import java.io.File
 import java.io.FileNotFoundException
 
@@ -30,9 +29,10 @@ internal class ConfigProvider(
         val releasePhase = getReleasePhaseConfig()
         val credentialsConfig = getCredentialsConfig()
         val releaseNotes = getReleaseNotesConfig()
-        val seoTags: List<SeoTag> = cli.seoTags ?: extension.seoTags
+        val seoTagIds: List<Int> = cli.seoTagIds ?: extension.seoTagIds
         val minAndroidVersion = cli.minAndroidVersion ?: extension.minAndroidVersion
         val developerContacts = getDeveloperContactsConfig()
+        val appType = cli.appType ?: extension.appType
 
         val artifactFile = getBuildFile(customBuildFilePath, artifactFormat)
         val artifactFileExtension = artifactFile.extension
@@ -55,9 +55,10 @@ internal class ConfigProvider(
             releasePhase = releasePhase,
             releaseNotes = releaseNotes,
             applicationId = applicationId,
-            seoTags = seoTags,
+            seoTagIds = seoTagIds,
             minAndroidVersion = minAndroidVersion,
             developerContacts = developerContacts,
+            appType = appType,
         )
     }